autotest: wifi: Add network_WiFi_BT_AntennaCoex
b:79233533 generated a large number of "wifi doesn't work until reboot"
feedback reports. The root cause was wifi not being able to
initialize/calibrate because BT was holding the shared antenna hostage.
Prevent this with a simple test that tests various sequences of wifi
off/on mixed with bt off/on. Note that there is no graceful failure with
this test - if the wifi interface is lost, shill decides to die with
LOG(FATAL), following which dhcpcd also kills itself (because no shill..) and
that breaks our ethernet control connection to the DUT. The fix for that
situation should be tackled in chromium:698457.

CQ-DEPEND=CL:1187837
BUG=b:79233533, chromium:876920
TEST=Ran this test on eve, samus, nocturne and bobba.
Removed the fix (CL:1162876) on 4.4 kernel and tested that this test borks
the eve DUT without the CL added.

Change-Id: I127de09881462c236ad07635fc1a8b841ecdc8cb
Reviewed-on: https://chromium-review.googlesource.com/1187564
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>