network_CheckCriticalProcesses: Add powerd and dhcpcd
Not sure why dhcpcd was omitted in the original list. If we are
connected to a DUT, it must be via wifi, ethernet or something
that most likely uses dhcpcd to get an IP. So checking for dhcpcd
being up is a good idea.
powerd was not in the original list of requirements, but with
convertible/detachables, it is necessary to reduce wifi transmit
power (being the daemon that calls set_wifi_transmit_power).

BUG=None
TEST=Ran the updated test on samus.
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>

Change-Id: Ib5fd467078f3f8f3c0aa7802a53494c38341f629
Reviewed-on: https://chromium-review.googlesource.com/1187898
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>