autotest: network_WiFi_AttenuatorPerf - Fix multiple issues
Fixing 3 issues that we have see on the attenuated (grover) setups in
the lab.
1) Unhandled timeout when opening telnet connection (crbug.com/854377)
2) TypeError for signal level. output_perf_value expects string or
number but wifi_signal_level returns None if signal is not detected.
This usually happens on runs with high attenuation (crbug.com/854378)
3) Telnet commands fails to return valid data. We use to see this error
sometimes prior to Aug 10 but since then failures resulting from this
has spiked. The issue seems to be with telnet connection that we create
though the connection remains open, we stop receiving expected data
using telnet expect. This seems to get worse with large number of
telnet calls. Closing and reopening telnet connection fixes the issue.
(crbug.com/876945)

BUG=chromium:854377, chromium:854378, chromium:876945
TEST=Tested against eve in grover setup in wifi lab.

Change-Id: Iccb711c59f29a216ea6a936c7757b4048fd0cc1c
Reviewed-on: https://chromium-review.googlesource.com/1188984
Commit-Ready: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>