autotest: Add server test to run cr50 vNVRAM test.
The cr50 vNVRAM test needs a cleared TPM to run. This test is a
wrapper around the client test which clears the TPM, so that the
client test can run. The TPM is cleared again after the test runs
to ensure no state from the test is left in the TPM.

BUG=b:112385847
TEST=test_that .. firmware_Cr50VirtualNVRamServer

Change-Id: I79e65b03933764d247d4a2e505b1331b6078d332
Reviewed-on: https://chromium-review.googlesource.com/1189252
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>