autotest: gs cache: truncate too long exception message
The response of some gs cache may includes long list of file names. This
is not readable when format it to an exception message.

This change truncate the long message to reasonab length.

Minor change: logging the call url/cmd to gs cache server.

BUG=chromium:875349
TEST=Ran unit tests.

Change-Id: Id212ae1b47172cd4a9f5ca206af05945380fbb86
Reviewed-on: https://chromium-review.googlesource.com/1194438
Commit-Ready: Congbin Guo <guocb@chromium.org>
Tested-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>