autotest: tko: connect using unix socket
In a previous change, https://chromium-review.googlesource.com/1164064,
we added a new configuration in shadow_config.ini, global_db_socket, and
return it as a new entry in the settings dict. It works well, however,
it breaks code using django.

Django only uses settings of 'HOST'. It can be either a real host name,
or a unix socket if it starts with '/'.

BUG=chromium:868052
TEST=Ran below code in python interpreter on guocb-dev-autotest2.cbf

    1) Verified TKO connection works:
    import common
    from autotest_lib.tko import db
    sql = db.db()

    2) Also verified the Django connection works.

    import common
    import os
    os.environ.setdefault('DJANGO_SETTINGS_MODULE',
                          'autotest_lib.frontend.settings')
    from dango.db import connections

    c = connections['global'].cursor()

Change-Id: I4bca254ae63591536a900457ebcfbc71493af00e
Reviewed-on: https://chromium-review.googlesource.com/1213781
Commit-Ready: Congbin Guo <guocb@chromium.org>
Tested-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>