power_CPUFreq: Accomodate acpi-cpufreq driver.
This test's charter was to enter cpufreq userspace governor and then
verify that 'available frequencies' could be set.  With many of
today's modern processors targeting an exact frequency is simply not
possible.

In the case of 'acpi-cpufreq' driver while we can set & get the
frequency however the hardware doesn't necessarily honor that value.

For that reason, I've modified test to simply go through the motions
of setting the various operating points and logging the data without
causing a failure if set & get frequencies don't match.

CL also did a bit of refactoring to:
- simplify save/restore logic
- unify frequency compare to accomodate crbug.com/848309 for both test
  modes (parallel & series)
- expose more debug logging to see values

BUG=b:113558496
TEST=power_CPUFreq passes on careena & spring

Change-Id: I3011ec308dda15d0dd276fdb6c632b509c4a1872
Reviewed-on: https://chromium-review.googlesource.com/1213773
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>