power_status: fix exception handling
Sometimes we read invalid state from the battery. To handle this, we
have a retry loop on reading this state. When it fails we print out
the contents of the battery state object, but remaining_time isn't
set at this point. This caused another exception from accessing a
field that did not exist in the object when we try to print it out
in exception handling. This fixes that by making sure all fields
are set before trying to raise an exception.

BUG=None
TEST=power_LoadTest

Change-Id: Idae39e21b21fe1776bc0f7341ff386bf009a2bb3
Reviewed-on: https://chromium-review.googlesource.com/1220607
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>