[autotest] Fix DUT availability check for run_suite.
The check for "Are there DUTs available for this suite" was failing to
account for suites that requested testing a against a model in addition
to a board and pool.  That resulted in wrong answers in multiple ways.

This changes RPCHelper.check_dut_availability() to account for all
of the dependencies that may have been requested in a given run_suite
call, and not just the board and pool dependencies.

Additionally, this fixes RPCHelper.diagnose_pool() to account for the
same constraints.

BUG=chromium:876121
TEST=Unit tests

Change-Id: Id6b7cbcc760c7c3221e9f4d382d2804b3a8fde92
Reviewed-on: https://chromium-review.googlesource.com/1222387
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>