[autotest] Fix idle DUT accounting in lab_inventory.
In inventory runs, DUTs that were locked when broken were being
accounted for as broken instead of idle, even if they hadn't done
work over the inventory time interval.

This changes handling so that DUTs that are locked will eventually
be accounted for as idle rather than broken.

This also adds a unit test to exercise the new behavior.

BUG=None
TEST=run manually with --debug.

Change-Id: Iab1ff65cbc3ab056392d8635e3360e5b579a0c82
Reviewed-on: https://chromium-review.googlesource.com/1236779
Tested-by: Richard Barnette <jrbarnette@google.com>
Reviewed-by: Congbin Guo <guocb@chromium.org>