autotest: Remove firmware_Cr50VirtualNVRam from all test suites.
This test needs to run after the TPM has been cleared, which requires
restarting the DUT. There is a wrapper test that handles this,
firmware_Cr50VirtualNVRamServer; it clears TPM and then test.
We should not try and run this test automatically as it will fail,
and is being run by the wrapper anyway.

BUG=b:115929325
TEST=none

Change-Id: I51e832e9a10683d5b95b9acc12112f87eddb24ed
Reviewed-on: https://chromium-review.googlesource.com/1233153
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>