[moblab] require part id, bug id for qual suites
Put a check that requires users to enter a part id and bug id for
storage qual and memory qual suites. The check is performed on the
client, and prevents users from submitting the suite until they
fill these fields in. Partners often forget to fill these in, and
they are required for buganizer notifications to work.

BUG=b:111728263
TEST=local moblab device

Change-Id: If871db5507148693936064c133950fa101821f0e
Reviewed-on: https://chromium-review.googlesource.com/1234934
Commit-Ready: Matt Mallett <mattmallett@chromium.org>
Tested-by: Matt Mallett <mattmallett@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>