tradefed_test: Simplified media file pushing.
The main purpose is to support caching those files for P CTS.

On N, CtsMediaStressTestCases was the only module that used these additional
files. However on P, 3 different modules depends on this media bundle, in 3
different ways. Rather than extending the existing code to the 3 modes, this
CL aims to let cts-traded to handle the variety of media file pushing. The
autotest side just downloads and caches the file content.

The intent of pushing media file by ourselves for ARC N in the past was to add
robustness to the otherwise weak and flaky lab network. Today, the lab network
allocation has been greatly improved and reliable. We haven't encounterd
corrupted media error for a long while. This part should now be safely skipped.

BUG=b:78434482
TEST=cheets_CTS_N.CtsMediaStressTestCases
TEST=cheets_CTS_N.7.1_r21.arm.CtsMediaStressTestCases
TEST=cheets_CTS_N.7.1_r21.arm.plan_follower
TEST=cheets_CTS_N.7.1_r21.arm.all
CQ-DEPEND=CL:*684908

Change-Id: I27923516b88e7c27441113925e5d3d030538eec2
Reviewed-on: https://chromium-review.googlesource.com/1235457
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>