[autotest] Restart servod in `deploy servo`.
This changes `deploy servo` to unconditionally restart `servod` on
the target servo.  This guarnatees that the cached GS URL in
`servod` gets cleared so that the USB stick can be re-imaged.  It
also makes `deploy servo` into an easy UI to force restarting a
`servod` job for an arbitrary DUT.

BUG=chromium:879295
TEST=run `deploy servo` locally

Change-Id: If3143b12ec5713db649ba0268459fb44262606d0
Reviewed-on: https://chromium-review.googlesource.com/1236784
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>