test_push: do not reverify DUTs again after staging test
Prior to this, we would call reverify after staging test, even is
staging test was a success and we printed the "staging success" message.
This leaves open the possibility for reverify to fail and the script to
return with a failure code even after we had declared success, which is
confusing. It also adds needless extra latency between when staging has
succeeded and when we actually promote our prod-next candidate (which is
done by the caller to test_push).

BUG=None
TEST=None

Change-Id: If654aaf0dc29c2be4ddf889229c06a913a1961d8
Reviewed-on: https://chromium-review.googlesource.com/1237079
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>