[autotest] site_linux_router: guard against empty syslog
If /var/log/messages delivers an empty line, .split() will deliver an
empty list, and so we hit this error:

      File "/usr/local/autotest/server/site_linux_router.py", line 194, in __setup
        self._log_start_timestamp = last_log_line.strip().split(None, 2)[0]
    IndexError: list index out of range

It's safer to use .partition(), which guarantees we'll get a result. And
if that result is just an empty string, replace it with a wildcard
regex.

BUG=chromium:889556
TEST=truncate /var/log/messages on router, run network_WiFi_* tests

Change-Id: I369649121658f41fb3deff4f9be32af54bf1f14a
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1246688
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>