[moblab] pass part_id, bug_id to memqual via test args
Currently part_id, bug_id are passed via suite args. This
complicates the pipeline, as it needs to delay processing
until the suite job has been uploaded. Pass the 2 variables
via test args, so that the memqual pipeline is not dependent
on the suite job.

BUG=b:92031054
TEST=run memqual on local device, verify that part_id, bug_id
are written to the keyval

Change-Id: I2dbd86c3db092bb2fea0cf76e3db93bd0a38ad2b
Reviewed-on: https://chromium-review.googlesource.com/1249615
Commit-Ready: Matt Mallett <mattmallett@chromium.org>
Tested-by: Matt Mallett <mattmallett@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>