[autotest] Add an optional cache layer to gs_offloader timestamps.
gs_offloader makes AFE calls to get the completed timestamp of
the jobs to be offloaded.

It makes at least 2x the number of calls it needs to

If jobs are staged for some days ( like on moblab ) locally these
AFE calls QPS get very large, up to 4800 qps.

Add in a cache for the result AFE call to get the timestamp.

TEST=ad hoc testing on moblab
BUG=chromium:890350

Change-Id: I53e5d0765671f945cf7bffc6b9bd65ba7ce80ff6
Reviewed-on: https://chromium-review.googlesource.com/1269735
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Matt Mallett <mattmallett@chromium.org>