security_x86Registers: Fix VMX-related failures.
This test might not be long for this world. It tests VMX state in the
corresponding MSR, and the BILD bit in firmware. Well, it turns out VMX
is being set in all the boards supporting Crostini, so that check is
not very useful anymore.

Remove the VMX check on Intel for now, but keep the BILD check, which
is still useful.

BUG=chromium:848428
TEST=Passes on soraka, caroline.

Change-Id: Icecba7b237dd7de99c4b1b590060e7e2752719af
Reviewed-on: https://chromium-review.googlesource.com/c/1259564
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>