set_device_mode: don't raise exception if not needed
`inject_powerd_input_event --code=tablet --value=0` fails on devices
that don't support tablet. Regarless of the value passed to `--value=`
it fails.

This CL catches the exception, and doesn't re-throw it if the desired
mode is clamshell, since it believes that the device is already in
clamshell mode.

BUG=b:116836237
TEST=test_that cros0 cheets_AppPositionTest_P
     Tested on Grunt. Test didn't receive the error.CmdError exception

Change-Id: If35e9f757ba00a17e15da728f92367d2797f33aa
Reviewed-on: https://chromium-review.googlesource.com/1260184
Commit-Ready: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Garfield Tan <xutan@chromium.org>