autotest: Perform cold reboot on test device to enable VM
There's an issue where some devices need to undergo a cold reboot in
order to reset the EC so a bit can get flipped to enable the VM. This
checks if the device supports VM functionality by testing the existence
of /usr/bin/vm_concierge and then if the cold reboot is needed by the
lack of /dev/kvm. If both those are true it uses ectool to do a cold
reboot.

BUG=chromium:892782
TEST=Manually verified on locked DUT

Change-Id: I5d1994d896814e874f6f5659263d90b63e08d970
Reviewed-on: https://chromium-review.googlesource.com/1262395
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>