autotest: reduce default test timeout from 24hr to 6hr
Shards are accumulating a large number of backlogged jobs, most of which
are doomed to never run, but which still contribute a lot of shard rpc
and scheduler overhead.

Reduce the default timeout from 24hr to 6hr to reduce this backlog.

BUG=chromium:894197
TEST=None

Change-Id: I561cb1e3586c1a5156402c53b6544c2fc05f041b
Reviewed-on: https://chromium-review.googlesource.com/1274395
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>