video_PlaybackPerf: Fail if an error occurs in HW playback
Originally, video_PlaybackPerf doesn't check an error status of VDA when HW
playback is used. There is case that playback is fallback to SW when VDA causes
an eror. In the case, SW playback performance values are reported as HW playback
ones.
This fails video_PlaybackPerf if an error in VDA occurs.

BUG=None
TEST=video_PlaybackPerf on nyan_big and veyron_minnie

Change-Id: I437e086809db734a789fa4b5ee7ecba7ba3bc1b1
Reviewed-on: https://chromium-review.googlesource.com/1278605
Commit-Ready: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@google.com>