power_LoadTest: skip web_page domains for keyvals
After introducing checkpoint logic for each website, the keyval file is
blowing up with detailed stats about each website. The keyval file is
not the place for that. This change skips all checkpoint data for
domains that contain [web_page_] in their name as its the prefix used
for specific web page data. The raw data is still preserved & stored.
If and when there are more usecases to skip domains besides web-pages,
then a more generalizable domain-skipping system (either in PLT, or in
the calc() function) might be appropiate, but that's not the case yet.

BUG=b:117680074
TEST=test_that ... power_LoadTest.fast

Change-Id: Ib49433809960736015ce67b21190bedc89933ac2
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1280482
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>