servo: set usb stick to poin to host before probing the device
This is an autotest-side fix to the issue described in the BUG
where devices are failing due to the new probing logic not explictly
pointing the usb stick towards the host, and thus not finding the usb
stick block device file.

Once servod is updated this change will not be needed anymore, but once
servod changes have all landed, this entire file needs to be updated
anyways. This change might alleviate some lab issues now, before a
servod fix can land.

BUG=chromium:895705
TEST=None yet

Change-Id: I49034200f8daa399cb3d8f4ba324b55b43689231
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1282954
Tested-by: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>