Update firmware_LockedME to skip the test on AMD devices.
AMD devices don't have an Intel ME, so the firmware_LockedME test
shouldn't be run on AMD devices. If the firmware_LockedME test is
running on an AMD device, it will skip the test and print a
message saying that the test is not applicable for that device.

BUG=b:117672700
TEST=On Careena, test_that --board=grunt {DUT_IP}
f:.*firmware_LockedME/control
TEST=On Bobba, test_that --board=octopus {DUT_IP}
f:.*firmware_LockedME/control

Change-Id: I1bbf76ce201f4e679d60c317c50352ee2598e746
Reviewed-on: https://chromium-review.googlesource.com/1287120
Commit-Ready: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>