graphics_Idle: Check PSR dynamically
At the moment, PSR is only checked on two boards,
Samus and Eve. This will check if the board is an
Intel board first, and if so, what kernel version
it's running. Based on the kernel version, PSR
will have different forms of active/inactive states.

If PSR is not enabled or if the board is not an Intel
device, the subtest is skipped and will count as
a pass.

BUG=None
TEST=test_that --board=<BOARD> <IP ADDRESS> graphics_Idle

Change-Id: I620b62edcde382f174a7c3867672d0708d5d8aa4
Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1295549
Commit-Ready: Casey G Bowman <casey.g.bowman@intel.com>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>