crash: overhaul report file processing
The current code hardcodes each file type that we know might be part
of a crash report which makes it hard to add more -- it requires a
good amount of copy & paste.

Lets rewrite the logic to have a set for required & optional file
types in each crash report, and have a generic loop to gather all
the files before we detect missing/unknown files.

Now if we want to add more required or optional files, we only have
to update the two class sets.

BUG=chromium:889552
TEST=`test_that ... logging_UserCrash` still passes

Change-Id: I8bbebdc8be607c58e130dfdc710ed2584b793d94
Reviewed-on: https://chromium-review.googlesource.com/1297192
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>