[autotest] network_WiFi_PTK: handle floats for loss percentage
Ping loss percentages can be floats -- they already are for the Mac OS
implementation, and iputils is starting to use them too (in part to
prevent erroneous values when near 0 or 100%).

So don't print them with %d, and add a float parser in the Linux ping
parser.

Now instead of 43.3333 showing up as this:

  Ping loss percentage: 3333.

We get:

  Ping loss percentage: 43.3333.

BUG=none
TEST=`network_WiFi_PTK` with a bad Wifi firmware, logs correct
     percentage

Change-Id: I233e3d5a0acfc7e64630c92677fe2e150a1d42e4
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1300597
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>