bluetooth: base64 to encode/decode non-ASCII data
When a DUT scanning devices nearby, it is possible that there
are non-ASCII data in the device properties, e.g, ServiceData
and ManufacturerData. While yaml is handy to serialize such
data, yaml is unfortunately not available on some test server
containers. It is possible to install yaml at the cost of
extra size and installation time which is what we would like
to avoid. Hence, it makes sense to work around using the
built-in json module.

The get_devices() is a list bluez dbus device objects. A recursive
function is implemented as utils to apply base64 encode and decode
methods on the device list. It calls recursively on sequences
and dictionaries, and handles the basic types with the desired
methods.

Another culprit of json is that json cannot handle non-string
keys in a dictionary. It converts for example a number into
strings. We need to fix this culprit when applying base64 decoding
method.

Some debug messages has been added to print the data during
encode/decode.

CQ-DEPEND=CL:1297878
BUG=chromium:879758
TEST=The following test should pass.
(cros) $ test_that --args="tester=${TESTER_IP}" "$DUT_IP"
bluetooth_Sanity_LEDiscovery

There should be no errors like this:
  'utf8' codec can't decode bytes in position 2-3: invalid continuation byte

Change-Id: I81f9445a0227f7c74b2d07cfa9003fa4e0afcf0b
Reviewed-on: https://chromium-review.googlesource.com/1301753
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>