bluetooth: use pickle to replace json
When a DUT scanning devices nearby, it is possible that there
are binary values in the dbus.ByteArray in the device properties,
e.g, ServiceData and ManufacturerData. While yaml is handy to
serialize such data, yaml is unfortunately not available on some
test server containers. It is possible to install yaml at the cost
of extra size and installation time which is what we would like
to avoid. Hence, it makes sense to work around using the standard
pickle module.

Note that the xmlrpclib.Binary class is needed to handle possible
non-XML characters.

The original json module is not used for two reasons:
- The json module cannot handle arbitary binary values in
  dbus.ByteArray. A much more complicated recursive function is needed
  to handle that.
- The json module cannot handle numeric key values in a dictionary .

CQ-DEPEND=CL:1297878
BUG=chromium:879758
TEST=The following test should pass.
(cros) $ test_that --args="tester=${TESTER_IP}" "$DUT_IP"
bluetooth_Sanity_LEDiscovery

There should be no errors like this:
  'utf8' codec can't decode bytes in position 2-3: invalid continuation byte

Change-Id: I138998f381f4e8aed3d3d0a11502936f240b19f0
Reviewed-on: https://chromium-review.googlesource.com/1304453
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>