security_ModuleLocking: Look in the right directory for modules
The module locking test was trying to find the kernel modules with:

  ko = utils.system_output("find /lib/modules -name '%s.ko'" % (module))
  return ko.splitlines()[0]

If you run this test on a machine where you've previously run
update_kernel it's possible that you could end up with the wrong
modules.  Let's use $(uname -r) to find the modules directory which
should avoid this problem.

BUG=b:118505449
TEST=test_that IP security_ModuleLocking

Change-Id: I9597d4fb5cf2bd9d34f8b2a371b97e179c4eb426
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1308740
Reviewed-by: Philip Chen <philipchen@chromium.org>