autotest: Make tast.py read run_error.txt.
Make the tast.py server test read a run_error.txt file that
is written by "tast run" when the entire test run is aborted
(e.g. due to the SSH connection to the DUT being lost). This
message is then reported as an error for any incomplete
tests, as it provides a better clue to the underlying
problem than "Test did not finish".

BUG=chromium:895716
TEST=added unit test; also killed tests on DUT while using
     test_that to run tast.informational and checked that
     the run error was reported for the last test in
     status.log

Change-Id: I90dc587e193cf5e8826991401f9c487020bcf1cb
Reviewed-on: https://chromium-review.googlesource.com/1308747
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>