power_LoadTest: Remove individual CPU/GPU freq from keyval
CPUStatsLogger logs the percentage of time spent in each
frequency for every sampling interval. This leads to a lot
of frequencies that will be its own section in result keyval.

This CL removes that data from keyval as the weight average
of CPU/GPU frequency is already in the keyval to make the
keyval file size manageable.

BUG=b:117680074
TEST=ran PLT.fast, keyval size is human readable now.

Change-Id: I26fdcbfa2b28b72058516ad440dd095d6abf0c99
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1312275
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>