re-revert "autotest: remove private values from RPM config""
Revert "autotest: remove private values from RPM config"

This reverts commit f3538424f8a2a3d5db30b651859d025c7a103161.
This also reverts 7690869d37f99d85f7c296972a5cd858cf90d269, which is the
most-recent re-landing of above.

There appears to be some issue in in production with this. The passwords that
reverting this reveals are not high value, and are available anyway in the repo
history.

BUG=chromium:866762, chromium:821300
TEST=None

Original change's description:
> autotest: remove private values from RPM config
>
> They're deployed in puppet, now.
>
> BUG=chromium:821300
> TEST=unit tests
>
> Change-Id: I72f6f91be9b96714254c846d20dc29e20f2a9ed8
> Reviewed-on: https://chromium-review.googlesource.com/1024253
> Commit-Ready: Jacob Kopczynski <jkop@chromium.org>
> Tested-by: Jacob Kopczynski <jkop@chromium.org>
> Reviewed-by: Jacob Kopczynski <jkop@chromium.org>

Change-Id: I4a543d921084aea3520466d412fea06956f23bd9
Reviewed-on: https://chromium-review.googlesource.com/c/1313662
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Alex Zamorzaev <zamorzaev@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>