Enable rlz guest mode ping test.
This test was disabled as it didn't work previously. Now we need get the
device 'locked' for rlz by first logging in as a real user and then
checking the ping in guest mode.

I have also added a timeout of 5 minutes so the test doesn't hog a
device if something goes wrong.

BUG=None
TEST=test_that <ip> rlz_CheckPing
TEST=test_that <ip> rlz_CheckPing.guest

Change-Id: If9cd3839eae20dd0dcb307c45b4e22c49f67727e
Reviewed-on: https://chromium-review.googlesource.com/1327621
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Stephen Lin <stephenlin@google.com>