retry.py: add exponential backoff
In server side autotest power_ServoV4ChargeWrapper, we would like to be
able to send command to Servo v4 to control whether it is a power source
or sink for the DUT. After sending the command, we would like to check
that the command we send is actually effective, by querying the power
role of Servo v4, and by checking the DUT power status. It usually take
around 1 second for the change to take effect, so we need a retry
decorator to keep querying until success or timeout. Exponential backoff
in the retry decorator helps us reduce the number of times to retry, but
still responds fast if Servo v4 command succeeds quickly. Thus, we are
adding the option to exponentially increase delay time, while
preserving the randomization.

BUG=b:117681170
TEST=test power_ServoV4ChargeWrapper modified in a later CL
test_that <DUT ip> power_ServoV4ChargeWrapper --args 'test=power_Dummy' \
--autotest_dir ~/trunk/src/third_party/autotest/files/

Change-Id: I6aa84e864aa7507070f91ca0ac289dab23bddf82
Reviewed-on: https://chromium-review.googlesource.com/1345457
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>