firmware_RecoveryCacheBootKeys: Make test accomodate EFS
When the EC is in RO and we do a recovery boot, it causes the EC to do
a soft reboot, which will in turn cause a PMIC reset (double reboot).
This causes us to lose the request to do MRC training for this test.
The solution is to make sure that the EC is in RW before doing a
recovery boot.  Also added extra reboot to cleanup function in case we
get an error while in recovery mode.

BUG=b:73293232
BRANCH=None
TEST=run firmware_RecoveryCacheBootKeys FAFT test

Change-Id: Ie6e530767ad71ff1420989e01bee7abe8ec10dd0
Signed-off-by: Shelley Chen <shchen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1356309
Commit-Ready: Shelley Chen <shchen@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>