[autotest] network_WiFi_ChannelScanDwellTime: correct list checks
CL:1097984 incorrectly refactored our list-collation logic. Today, it's
not possible to hit the 'Failed to find any BSS' code, since the only
TimeoutError is when _bss_list is None (which is different than []).

Instead, we get this, when the empty list makes its way into logic that
didn't expect it:

  File "/usr/local/autotest/server/site_tests/network_WiFi_ChannelScanDwellTime/network_WiFi_ChannelScanDwellTime.py", line 82, in _get_dwell_time
    index_diff = ssid_index[-1] - ssid_index[0]
IndexError: list index out of range

I picked a semi-related BUG=, because we're not seeing that symptom any
more, and are instead seeing this Error. We still should probably figure
out why we fail the test.

BUG=chromium:628784
TEST=network_WiFi_ChannelScanDwellTime

Change-Id: Ie458ffb609ebdf8940ec793f782befd293d2c4a2
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1357253
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>