platform_ExternalUSBStress: remove redundat op
This test looks to change USB key direction by directly using a servo
v3 control, right after changing the direction with a portable method.
So, this gets rid of the direct manipulation in favor of the usage of
the portable method (so should have no change to what the test does).

BUG=b:79940760
TEST=ran on cyan, observed that the test got further than the key
direction change

Change-Id: I9b1183da5765163e24d9d43e7e76afdaf8754ee3
Reviewed-on: https://chromium-review.googlesource.com/1364510
Commit-Ready: Kevin Shelton <kmshelton@chromium.org>
Tested-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>