autotest platform_PrinterPpds: upgrade to version 3
This is a new release of the autotest platform_PrinterPpds. It was extended by
a parser of /var/log/messages. The parser can extract CUPS logs and interpret
them, what allows for better detection of internal errors. Tests can be also
run in "debug" mode. In this mode CUPS log level is increased and the parser
extracts whole printing pipeline as a bash script. This script can be run
outside CUPS for debug purposes.

BUG=chromium:898734
TEST=Tested on elm and nautilus

Change-Id: I7de2fe36d46b0c72c55767910d4ebe2f4e1d467c
Reviewed-on: https://chromium-review.googlesource.com/1369116
Commit-Ready: Piotr Pawliczek <pawliczek@chromium.org>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>