[autotest][cfm] Makes enrollment flow less flaky
The Hangouts App might restart a couple of time before reaching the OOB
page. We extend the timeout after enrolling the device to be more resiliant
to such restarts.

Why not use a wait instead? In order to do that we need to restart the app
ourselves to obtain a handle to chrome. Depending on when the regular restarts
occur, we might restart during the `stabilization` time and loose the handle anyways.

TEST=local device and lab device
BUG=b:120766786

Change-Id: Iaf20aefed512fdc99ac66524070a387250d02586
Reviewed-on: https://chromium-review.googlesource.com/c/1370050
Tested-by: Denis Tosic <dtosic@google.com>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Denis Tosic <dtosic@google.com>