[autotest][cfm] Adds an action for joining a meeting with bots
This enables us to have multiuser meetings in configurable CfM tests.
Also implements a simple scenario where we are in a meeting for a longer
period of time - for now 20min, but we will increase it once we confirm
the test runs happily and stable.

TEST=local device
BUG=b:119806655

Change-Id: Iccc4a1bcd8169eb818898b8b4f37a00b9c293e73
Reviewed-on: https://chromium-review.googlesource.com/1375073
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Denis Tosic <dtosic@google.com>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Claes Malmnäs <malmnas@google.com>
Reviewed-by: Denis Tosic <dtosic@google.com>