network_WiFi_RoamFT: add check for supplicant roaming
Some devices are failing this autotest because they support driver
based roaming. This means that the roaming decision is outside of
software control and shouldn't be tested with this autotest. Add a
check to require that the device does not support driver based
roaming.

BUG=chromium:907656
TEST=Ran tests on chromeos15-row1-rack8-host2.cros which supports
     driver based roaming and got TestNA. Ran tests on local caroline
     and everything passed.

Change-Id: I305d0dd6d64714d1434ed7eeecb5899a3a5390d4
Reviewed-on: https://chromium-review.googlesource.com/1376889
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>